Fixes for bugs: http://bugzilla.vyatta.com/show_bug.cgi?id=3549 and  http://bugzilla.vyatta.com/show_bug.cgi?id=3552 This fix involves removing multiple spaces between args on the kernel command line.
This fixes the symptom, but the root cause is an uninitialized variable in the grub-pc code,
so there is absolutely no guarrantee that this problem will not arrise again.  The correct fix
is either to return to grub (V1) or to have the developers find and fix the unitialized memory.