Allow any number of tunnel devices
The CLI should not be enforcing arbitrary restrictions on the number of
devices. The limit should be based on the resources of the system and
only the kernel would know that. So let the user try and make 10000
tunnel devices, maybe its on a 2000 core machine with 64G of memory.
Any failure will show up when the attempt is made to create the tunnel
device.