Bugfix: 3744
When a new member is added to a RAID group that holds the root filesystem,
we need to re-install grub so that the new disk will be bootable.  But this
can only be done after the RAID set has completed rebuilding.  Added mechanism
that uses the event notification infrastructure of "mdadm" to trigger
the re-installation of grub after the rebuild completes.