Bugfix 4490: Mark root partition as "bootable" in the master boot record (MBR).
The MBR provides a flag on each of the partitions to indicate whether it
is "bootable" or not.  Most BIOS implementations ignore this flag and
will boot from a partition whether or not this flag is set.  Some BIOS
implementations, however, refuse to boot from a partition that is not marked
as "bootable".  This fix marks the partition that holds the root filesystem
as "bootable" so that it can be booted by either type of BIOS.