Bugfix 4700:  Move processing to "end" field so that param modifications are handled.
In order to correctly handle cases where parameters are changed after a repository
configuration has been set up, we need to do the work in an "end" field instead
of a "create" field.  The "create" field is only executed when the node itself
is created, not when nodes below it are modified.