Bugfix 4892:  Add a config parameter to disable IPv6
Added the parameter "system ipv6 disable".  If set, the module
parameter "disable_ipv6=1" will be passed to the IPv6 module.  This
inhibits assignment of IPv6 addresses on any interfaces, effectively
preventing any IPv6 communication.  But it leaves the IPv6 module
loaded so that applications that depend upon its existence will not
fail.