Changes related to co-ordination between kernel parameters and radvd.
First, we need to set the global IPv6 parameter under "all" when
"disable-forwarding" is deleted because this is the parameter that
actually controls whether the stack will forward IPv6 packets.

Second, if router advertisements were configured while global IPv6
forwarding was disabled, we need to re-start the daemon when global
IPv6 forwarding is re-enabled.