Bugfix 5240: Prevent re-installation of same version on running systems
Previously, it was possible in some cases to "add" a system image that
would over-write the running image.  This change fixes that by getting
the version string of the running system from the kernel command line
instead of the Vyatta version file.  This string represents the actual
directory name that holds the image, so is more reliable.

Also previously, an attempt to "add" a system image that had the same
version string as another image installed on the system (but was not
the running image) would always succeed.  This would completely
over-write that image.  Because of this destructive behavior, it
deserves a confirmation question.  This change prompts the user to see
if they really want to do that before proceeding.