Bugfix 5120: Use biosdevname to generate the canonical name for an interface.
Network interfaces are discovered at boot time in non-deterministic
order, so their names are non-deterministic.  The vyatta_net_name
script is responsible for ensuring that network interface names are
changed to match their Vyatta config file entry when a match based on
MAC address is found.  With this change, when a match is not found, we
use the biosdevname tool to to find the canonical name for an
interface, then change the interface's name to that name.  This
ensures that NICs are named in a deterministic and systemmatic fashion.