Bugfix 5360: Get name of image from correct source.
The name of a system image is supposed to match the Vyatta version
string.  Previously, the install-image script determined the name of
the image being installed by parsing a field in the Vyatta version
file on the image being installed.  But the format of that file has
changed, and may change again the future, so it is not a good source
of the Vyatta version string.  The most accurate way to determine the
Vyatta version string is to look up the debian package version string
associated with the vyatta-version package on the image being
installed.  This changes the install-image script to get the system
image name that way.
(cherry picked from commit e1864411c77e688b1b67de40441550b1f1eba055)