Bugfix 5512: Have grub.cfg use the vmlinuz and initrd.img symlinks if available
The /boot directory is populated with vmlinuz and initrd.img files
whose names are suffixed with the kernel version string.  A vmlinuz
and initrd.img symlink points to the "current" version of each file.
It is simpler to have the grub.cfg file use the symlink, since then
the kernel to boot can be changed by just adjusting the symlinks.