Bugfix 6085: Use "-x" flag instead of "-r" to kill the old dhclient process
The "-r" flag to "dhclient" kills the old dhclient daemon, but also
tries to release the lease on the server.  Unfortunately, the program
will block forever if a lease had not actually been obtained.  The
"-x" flag does all the same things as "-r", but doesn't try to release
the lease on the server, and hence doesn't block in the case where the
lease had not been obtained.