Bugfix 7502: Get "add system image" to work on Xen
This fix solves the problem with indirection.  We introduce a symlink
in /live/image/boot pointing to the current default system image.  The
pygrub config file in /boot/grub/menu.lst points to the kernel and
initrd files in the default image by using this symlink in the
pathname.

Every time we add, delete or change the default system image, we need
to change this symlink to point to the new default image.