Use hash map of addresses to interfaces
Other part of fix for Bz 8171
Create a hash of address => interfaces in configuration
and use it during address commit check.

Remove unused creation of list of current addresses.
This might speed boot with large number of interfaces.

Since the hash is made once per interface rather than
multiple times, it might be faster because of that as well.