Fixed parted command for root partition alignment.
The current parted command does not align the root partition as desired
by the `--align optimal` option. This is evident from this parted message:

"Warning: The resulting partition is not properly aligned for best performance."

Instead of setting the ROOT_OFFSET=64S or ROOT_OFFSET=1024S (> 10GB) we should
set it to 0%. This will result in optimal alignment regardless of disk size.

Source: https://unix.stackexchange.com/a/49274