vyatta-cfg-system: allow interfaces with dhcp addresses to be deleted
Update the checks in scripts/vyatta-address to allow the deletetion of
an interface that has a DHCP address.  The patch put in place for
Bug #259 (http://bugzilla.vyos.net/show_bug.cgi?id=259) checks that an
interface exists before preceding with the delete, but this chokes when
supplied with the address 'dhcp'.

This patch checks if the supplied interface is a DHCP address, before
extracting the IP address and then using that address for the check.

Bug #305 http://bugzilla.vyos.net/show_bug.cgi?id=305