vyatta-cfg-system: fix for 'user vyos doesn't exist' on password reset
If you delete the vyos user from the system and then go into password
reset mode on reboot, the configuration entry for the vyos user is
recreated but the command returns 'usermod: user 'vyos' does not exist'.
Despite this message the user is created correctly on a full reboot,
just not at the point the command is run.

Bug #342 http://bugzilla.vyos.net/show_bug.cgi?id=342