vyatta-cfg-system: redesign the layout and updating of /etc/resolv.conf
Following the documentation, /etc/resolv.conf should only have comments
inserted that start with # at the beginning of a line, comments in
other locations are not supported.  The exisiting method of tracking
changes in this file (done by both vyatta-system-nameservers and
vyatta_update_resolv.pl) relies on this unsupported feature.

This commit restructures the updates so they use comments on the
preceding line to track changes, along with updating the old style
comments to the new ones when the relevant scripts are called.

This also fixes a previous issue (which was reverted) where IPv6
nameservers worked incorrectly when the /etc/resolv.conf line ended
with a comment.

Bug #486 http://bugzilla.vyos.net/show_bug.cgi?id=486