vyatta-cfg-system: redesign the layout and updating of ntp
When NTP is configured on the system, it not only acts as a client
for the configured servers, but also as a server.  Although the server
is only available as a time source, it still represents a service that
the user hasn't specifically enabled.

This commit sets the default configuration of NTP to disallow all
external access, so the system acts purely as a client by default, and
also introduces the ability to configure which addresses / subnets are
allowed to both query and use it as a time source.  If the servers
configured are specified as host names, these are resolved to IP
addresses before being added to ntp.conf, with the same process carried
out after name server changes.

syntax 'set system ntp client address'

Bug #94 http://bugzilla.vyos.net/show_bug.cgi?id=94