vyatta-cfg-system: ensure grub is installed to the raid slave members
When the system is installed in a raid configuration, grub is installed
to the raid device instead of to the individual slave drives, resulting
in a failure to boot after installation.  This commit identifies the
slave drives and ensures grub is installed to all of them (rather than
just the primary), resulting in a system that will boot following a
primary drive failure.

Bug #106 http://bugzilla.vyos.net/show_bug.cgi?id=106