vyatta-cfg-system: update protocol, remove server and node wording
Update the default protocol used for the dslreports service in the
revised dynamic DNS system and remove the default server settings for
freedns as it's already defined in ddclient.  Uodate the wording in the
server node to indicate it's only needed for custom services.

Bug #507 http://bugzilla.vyos.net/show_bug.cgi?id=507