vyatta-cfg-system: missing encrypted-password breaks user config node
The fix for bug #557 now allows the encrypted-password Authentication
node to be deleted again, but this causes the config to fail on reboot,
leaving the user login node empty.

This fix checks for an empty / missing encrypted-password node on
reboot / config reload, either creating a new node or replacing the
existing one and assigning it the value of "!".  This has the same
effect as the fix for bug #336, allowing the user to be set as only
using RSA based logins.

Bug #573 http://bugzilla.vyos.net/show_bug.cgi?id=573