vyatta-cfg-system: rename ethernet offload option nodes
Updates to the changes committed in connection with Bug #365 to allow
a workaround for Xen based network issues.

https://github.com/vyos/vyatta-cfg-system/commit/7a5a8500091589f0ede3d5a1dbc8b6b4b3fabf1d

Rename the offload configuration node to offload-options, shorten the
child node names, rename offload-settings to offload-option and add a
more descriptive error message if an option fails to apply to the
interface.

Bug #612 http://bugzilla.vyos.net/show_bug.cgi?id=612