vyatta-cfg-system: prevent deletion of all authentication for user
Prevent the deletion of both the encrypted-password and public-keys
nodes for a user, ensuring that at least one of the methods is
available.  Also prevent the deletion of the public keys for a user
if the encrypted password is set to "!".

Correct fix this time, without introducing the inability to remove
configured users from the system.

Bug #574 http://bugzilla.vyos.net/show_bug.cgi?id=574