T1045: static route dhcp-interface: check for ip in response
Checks done to verify a dhcp-lease on the interface is updated to look for a
ip-address like response from dhclient before returning an error.

This fixes the error where a newline was enough to make the checks fail..

Updated to also make sure no newline is returned after the address to make
sure commands using adresses from this script is not cut in half by the newline,
and to be in line with the syntax on all errors. (no newline at the end)

(cherry picked from commit 8f4b8917d23b42d15b7f48eb3f814a56cc5a4343)