install-image: T2690: T2762: always run in default VRF unless specified
When you connect via SSH and run the "add system image" command, it is launched
inside the VRF you are currently connected to. This might become confusing as
the VRF context changes.

Change command to always run in "default" vrf unless "add system image" command
is invoked with the vrf option.