boot: T5739: Fixed password recovery
This change fixes a filter used for `sed` during password reset in the
`set_encrypted_password()` function.

Old logic: search between `user username {` and the first line with `}` does not
work for users with extra options in config, like public keys.

New logic: search between `user username {` and the first line with
`encrypted-password` which should cover all possible combinations, except
situations when a password is not presented in a configuration file.